requirements.txt
anyio==3.6.1
asgiref==3.5.2
autopep8==1.6.0
click==8.1.3
colorama==0.4.5
fastapi==0.78.0
h11==0.13.0
httptools==0.4.0
idna==3.3
mysql-connector-python==8.0.29
protobuf==4.21.1
pycodestyle==2.8.0
pydantic==1.9.1
python-dotenv==0.20.0
PyYAML==6.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing_extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
